[GRIND75] Valid Parentheses
I found this one interesting because I did the C++ solution first and
all of the "where are we on the stack" was handled by the STL. When
working on the Python and C it made me rethink what the STL's are doing.
Are they being lazy with deallocation of memory on the array pop?
Could it be more efficient to just clear the memory at the back and not
use .pop and only use push if we've filled the stacks(vec) space.

This all took about ~30 min just due to typing and looking up which STL
methods are available in which language.