[GRIND75] 04: buy sell stocks
This one always trips me up, it was really good to run it a bunch of times and
get the performance data.  It was interesting at a LeetCode level the how
performant some languages are, but more importantly how much better the solution
in time is better than other submission but still way slower than all the others.

In adition I didn't fight the burrow checker as much this time for Rust.  I had
to look up some Vec operations but those were straight forward.  I also didn't
expect to have to derefernce the items returned from the prices.iter().  The
compiler was super useful in this case.