[GRIND75] 05 Valid Palindrome
I liked this one as it was much more interesting than the normal palindrome
question where it only wanted Alpha Numeric Palidromes but would include
non-alphanum ascii characters through out.  I approached this in a weird
way in C on the first day and it took me longer than I wanted as I was
over complicating it. A couple days later I came back to do the rest of
C++ , Python3, and Rust.  I was able to come up with a more optimal and
simpler to read solution without implementing everything myself.  Rust was
the make trickiness due to the String not being a byte string and that
triaging took me a while.

Over all I really like this version of the question.